feat(release): sign and notarize macOS binaries (#73)

+++ internal/localipc/stat_darwin.go
@@ -5,7 +5,8 @@ package localipc
 import "golang.org/x/sys/unix"
 
 func unixStatDevice(stat unix.Stat_t) uint64 {
-	return uint64(stat.Dev)
+	// #nosec G115 -- dev_t is a signed 32-bit C type; preserve its raw identifier bits.
+	return uint64(uint32(stat.Dev))
 }
 
 func unixStatMode(stat unix.Stat_t) uint32 {
